pyproject.toml
[project]
name = "proveit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.0.0",
    "python-dotenv>=1.0.0",
]
